Added a conversionLmatrix(a,b,c,d,N) function. This computes the change of basis operator from the polynomials with jacobioperator(a,b) to the polynomials with jacobioperator(c,d). It's still based on a linear recurrence relation. The dependencies are still in an "arrow head shape".
Changed the spectralmeasures function so that if we only have continuous spectrum then we get a Chebyshev T series (weighted by (1-x^2)^-.5). It's a little slow at the moment because we're doing this in a really inefficient manner: compute the L matrix from Chebyshev T to jacobioperator(a,b) for FULL SIZE 10,000 by 10,000. This is stupid but will work for the moment while we work out the most efficient way to compute the entries of L. It appears to work as hoped i.e. the number of coefficients in the expansion are very few when we nearly have singularities at the end points. And when we go for the semicircle, we simple get a quadratic (1-x^2).